requirements.txt
streamlit
pyvis
rdflib
pydotplus
pyyaml
lxml
bs4
click==8.0.4
